api 추가, 태그입력화면 하나로 통일

+++ app/src/main/java/com/syncrown/arpang/network/model/RequestSaveEditorDto.kt
@@ -22,7 +22,7 @@ class RequestSaveEditorDto {
 
     @SerializedName("pixel_x")
     @Expose
-    var pixel_x: Double? = null
+    var pixel_x: Int? = null
 
     @SerializedName("share_hash_tag")
     @Expose
