+++ app/src/main/java/com/syncrown/arpang/network/model/ResponseStorageContentListDto.kt
@@ -23,7 +23,7 @@ class ResponseStorageContentListDto {
 
         @SerializedName("pixel_x")
         @Expose
-        var pixel_x: Double? = null
+        var pixel_x: Int? = null
 
         @SerializedName("save_ds")
         @Expose
